Patch script paths inside of phrases and scripts
When numen scripts are first installed, they are properly copied to the nix
store location in `/nix/store/` even though we only specify an absolute path of
`/etc/numen`. However, since the numen install process still thinks the scripts
are at the absolute path starting with `/etc/numen`, it incorrectly patches the
paths in the the phrases. This just manually corrects those paths as part of the
patch process.